DB-backend/myproj/requirements.txt
asgiref==3.7.2
Django==3.2
django-cors-headers==4.2.0
djangorestframework==3.14.0
mssql-django==1.3
pymssql==2.2.7
pyodbc==4.0.39
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
